fix crawler logic do on api server
