+++ src/api/build.gradle.kts
@@ -13,6 +13,8 @@ dependencies {
     runtimeOnly("io.jsonwebtoken:jjwt-jackson:0.10.7")
 
     implementation(project(":storage"))
+
+    implementation("org.jsoup:jsoup:1.15.3")
     implementation("io.grpc:grpc-protobuf:1.53.0")
     implementation("io.grpc:grpc-kotlin-stub:1.3.0")
     implementation("io.jsonwebtoken:jjwt-api:0.10.5")
