+++ src/api/src/main/kotlin/com/tenmm/tilserver/crawler/application/inbound/DoCrawlingUseCase.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.crawler.application.inbound
+
+import com.tenmm.tilserver.crawler.domain.Identifier
+import com.tenmm.tilserver.crawler.domain.Url
+
+interface DoCrawlingUseCase {
+    suspend fun invoke(url: Url, userIdentifier: Identifier): Identifier
+}
