+++ src/api/src/main/kotlin/com/tenmm/tilserver/crawler/application/outbound/DoCrawlingPort.kt
@@ -1,0 +1,12 @@
+package com.tenmm.tilserver.crawler.application.outbound
+
+import com.tenmm.tilserver.crawler.domain.CssSelectorInfo
+import com.tenmm.tilserver.crawler.domain.Post
+import com.tenmm.tilserver.crawler.domain.Url
+
+interface DoCrawlingPort {
+    fun crawling(
+        url: Url,
+        cssSelectorInfo: CssSelectorInfo,
+    ): Post
+}
