+++ src/api/src/main/kotlin/com/tenmm/tilserver/crawler/application/outbound/SaveCrawlingPort.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.crawler.application.outbound
+
+import com.tenmm.tilserver.crawler.domain.Identifier
+import com.tenmm.tilserver.crawler.domain.Post
+
+interface SaveCrawlingPort {
+    suspend fun saveCrawling(userIdentifier: Identifier, post: Post): Identifier
+}
