+++ src/api/src/main/kotlin/com/tenmm/tilserver/crawler/application/service/DoCrawlingService.kt
@@ -1,12 +1,12 @@
-package com.tenmm.crawler.post.application.service
+package com.tenmm.tilserver.crawler.application.service
 
-import com.tenmm.crawler.post.application.inbound.DoCrawlingUseCase
-import com.tenmm.crawler.post.application.outbound.DoCrawlingPort
-import com.tenmm.crawler.post.application.outbound.SaveCrawlingPort
-import com.tenmm.crawler.post.application.service.crawler.CrawlerMap
-import com.tenmm.crawler.post.domain.Identifier
-import com.tenmm.crawler.post.domain.Url
-import com.tenmm.crawler.util.UrlCheck
+import com.tenmm.tilserver.crawler.application.inbound.DoCrawlingUseCase
+import com.tenmm.tilserver.crawler.application.outbound.DoCrawlingPort
+import com.tenmm.tilserver.crawler.application.outbound.SaveCrawlingPort
+import com.tenmm.tilserver.crawler.application.service.crawler.CrawlerMap
+import com.tenmm.tilserver.crawler.domain.Identifier
+import com.tenmm.tilserver.crawler.domain.Url
+import com.tenmm.tilserver.crawler.util.UrlCheck
 import org.springframework.stereotype.Service
 
 @Service
