+++ src/api/src/main/kotlin/com/tenmm/tilserver/crawler/application/service/crawler/NaverCrawlingService.kt
@@ -1,6 +1,6 @@
-package com.tenmm.crawler.post.application.service.crawler
+package com.tenmm.tilserver.crawler.application.service.crawler
 
-class MediumCrawlingService : CrawlerMetaService() {
+class NaverCrawlingService : CrawlerMetaService() {
     override fun getTitleCssSelector(): String {
         return "meta[property=og:title]"
     }
@@ -14,6 +14,6 @@ class MediumCrawlingService : CrawlerMetaService() {
     }
 
     override fun getDateFormat(): String {
-        return "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
+        return "yyyy. MM. dd. HH:mm"
     }
 }
