--- src/crawler/src/main/kotlin/com/tenmm/crawler/post/application/inbound/DoCrawlingUseCase.kt
@@ -1,8 +1,0 @@
-package com.tenmm.crawler.post.application.inbound
-
-import com.tenmm.crawler.post.domain.Identifier
-import com.tenmm.crawler.post.domain.Url
-
-interface DoCrawlingUseCase {
-    suspend fun invoke(url: Url, userIdentifier: Identifier): Identifier
-}
