--- src/crawler/src/main/kotlin/com/tenmm/crawler/post/application/outbound/SaveCrawlingPort.kt
@@ -1,8 +1,0 @@
-package com.tenmm.crawler.post.application.outbound
-
-import com.tenmm.crawler.post.domain.Identifier
-import com.tenmm.crawler.post.domain.Post
-
-interface SaveCrawlingPort {
-    suspend fun saveCrawling(userIdentifier: Identifier, post: Post): Identifier
-}
